[feat]#12 서버통신 responsedto 만드는 중
