+++ app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/data/dto/response/ResponseBase.kt
@@ -1,0 +1,11 @@
+package kr.co.nottodo.presentation.toplevel.recommendation.data.dto.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseBase<T>(
+    val status: Int,
+    val success: Boolean,
+    val message: String,
+    val data: T,
+)
