+++ app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/data/dto/response/ResponseRecommendationCategoryListDto.kt
@@ -1,5 +1,6 @@
 package kr.co.nottodo.presentation.toplevel.recommendation.data.dto.response
 
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -9,7 +10,13 @@ data class ResponseRecommendationCategoryListDto(
     val message: String,
     val data: List<ResponserecommendationCategoryListDetailDto>,
 )
-//"status": 200,
-//"success": true,
-//"message": "추천 카테고리 조회 성공",
-//"data":
+{
+    @Serializable
+    data class ResponserecommendationCategoryListDetailDto(
+        val id: Int,
+        val name: String,
+        val image: String,
+        val activeImage: String,
+    )
+}
+
