+++ app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/data/dto/response/ResponseRecommendationCategorySituationDto.kt
@@ -1,0 +1,17 @@
+package kr.co.nottodo.presentation.toplevel.recommendation.data.dto.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseRecommendationCategorySituationDto(
+    val status: Int,
+    val success: Boolean,
+    val message: String,
+    val data: List<ResponseRecommendationCategorySituationDetailDto>,
+) {
+    @Serializable
+    data class ResponseRecommendationCategorySituationDetailDto(
+        val title: String,
+        val action: String,
+    )
+}
